refactor(product): change price type from int to double for better precision
- Updated ProductItem model to use double for price instead of int, enhancing accuracy in price representation.
- Adjusted related serialization and UI components to accommodate the new price type.
- Ensured consistent handling of price data across various screens and controllers.

+++ lib/domain/models/response/get_store_info_food_response.dart
@@ -41,7 +41,7 @@ abstract class ProductItem with _$ProductItem {
     @Default(0) int id,
     @Default('') String name,
     @JsonKey(name: 'image_url') @Default('') String imageUrl,
-    @Default(0) int price,
+    @Default(0.0) double price,
     @JsonKey(name: 'avg_rating') @Default(0.0) double avgRating,
     @JsonKey(name: 'rating_count') @Default(0) int ratingCount,
   }) = _ProductItem;
